Restructure project into two python packages using python uv
First of all, the project is split into multiple python packages using a
standard python package layout. The first package is
hull_parameterization which is considered a separate concept as it has
different use-cases with different dependencies as opposed to just the
c_shipgen part. Then, the rest of the project as it was structured
previously is put into c_shipgen package without many differences.

A pyproject.toml file is added for each of the packages, and is tested
to work with python uv. The hull_parameterization package is added as an
editable package in c_shipgen with a relative path. The jupyter
notebooks have also been tested for the import section for the new
package structure. Finally, the readme has been extended to show how the
jupyter notebooks can be used with the new structure.