Replace gen_stl with gen_trimesh
Trimesh is a more extensive mesh geometry library with mesh operations,
algorithms and import/export. This is a slightly larger library than
numpy-stl which it replaces, but should still be capable of the same
thing. However, the main advantage is that the gen_stl doesn't export in
the function anymore, and a trimesh.Trimesh object is instead returned
which can be exported using the export method of that object. This
should enable the user to do whatever they want with the mesh, and also
support their 3D mesh representation format of choise supported by
trimesh.