Add type hints to hull_parameterization
This should first of all add type hints to most of the methods of
HullParameterization.py, but some are left unchanged since I don't know
the types as of now. These are not verified with e.g. mypy yet, so they
are mostly guesses or type printouts from running the code. This was
mostly added to give some better understanding of what the code does.

Some cleanup is done with code comments in some triple quote comments
have been converted to standard comments. Some unnecessary comments have
been removed.

Some smaller typos have been fixed.

Input defaults have been fixed where default parameters are set to [].
Instead, the type is specified as optional with default as None, which
is set to [] in the method body. This is because setting [] as a
parameter default is an antipattern because it is instantiated as a
common object at class level instead of at function call.